Fix transformation matrix calculation in CameraXBasic
This CL fixes transformation matrix calculation in CameraXBasic. The
transformation matrix here is almost identical to the one in
Camera2Basic except that CameraXBasic tries to merge the 2 scaling steps
together (scale back to the original buffer and scale it down to just
cover the viewfinder). In it, it makes false assumptions about the
orientation of the preview buffer. This assumption propagates the code
written below, causing additional scaling issues on landscape devices
with sensor orientation equal to 0. We rewrite the code here completely
to address the aforementioned issues.

Test: Tested on a 2-in-1 Chromebook and Google Pixel 3 and verified that
CameraXBasic preview looks fine on both devices in landscape and
portrait.