Fix transformation matrix calculation and layout (#127)
The CL fixes the following issues:

- Transformation matrix calculation
When calculating transformation, we need to take sensor orientation into
consideration. The first step of the calculation tries to resize the
rectangle back to its original buffer dimensions. However it falsely
assumes that the buffer dimensions are (preview height, preview width)
(i.e., buffer is rotated by 90/270 degrees).
Another bug happens when scaling the rectangle down/up so that it just
covers the viewfinder. It falsely assumes that viewWidth is the longer
edge and viewHeight is the shorter edge. This causes the preview to be
potentially scaled up too large, leading to a "zoom-in" effect.
- Landscape layout
When texture view is able to fill most of the screen, the control layout
will be squished to a point the width becomes too small to make the
shutter button visible.

Test: Tested this on a Chromebook and Google Pixel 3