Fix aspect ratio issues with preview
It seems that ConstraintLayout resizes children so that they don't clip.
This creates problems for AutoFitSurfaceView. AutoFitSurfaceView
attempts to set the measured dimension to be "just bigger" than the
viewfinder and expect that the overflowing pixels would be clipped,
creating the desired "center-crop" effect.
This commit changes the parent layout to FrameLayout.

TEST=Deployed on Pixel 4 AVD and verify the aspect ratios look right in
the test scene.