Use MediaCodec API instead of MediaRecorder and add 10-bit support.
10-bit encoding is not available using the MediaRecorder API. Because of
this, we must use the MediaCodec API instead. This patch introduces an
EncoderWrapper for this and checks for the dynamic range profile to set
the color space / bit depth of the encoding.

Since MediaCodec does not support 10-bit encoding with textures, the
texture view activity is no longer a choice if the user picks an HDR
dynamic range profile, and all references to dynamic range have been
removed from its code.

Change-Id: I4ab27a192b2a5668d487d9386c2078fc79db4dcf