Camera2Video: Fix MediaRecorder issues.
MediaRecorder recordings were sideways in portrait orientation, and
would not correctly record the HardwarePipeline. The former was due to
the lack of an orientation hint to MediaRecorder and an incorrectly
sized Surface. The latter was due to the presentation time being set,
which is incompatible with MediaRecorder.

Passing the Surface to a temporary MediaRecorder will set its size
correctly, and passing an orientation hint will rotate the output video.
Removing presentation time will ensure that frames' presentation time
in the output video start at zero and increment by 1/FPS seconds for
each frame. onReady would also be signaled at the beginning of recording
for the software path, which has been replaced by onClose.

Change-Id: Ib6d578a408413440c2071245944fc161db783cd3